rm lnauth profile settings
